move on_focus tests to separate files
