+++ components/dash-table/tests/selenium/test_sizing.py
@@ -1,12 +1,7 @@
 import dash
-import pytest
 
 from utils import (
-    basic_modes,
     get_props,
-    generate_mock_data,
-    generate_markdown_mock_data,
-    generate_mixed_markdown_data,
 )
 
 from dash.dependencies import Input, Output
@@ -324,30 +319,3 @@ def on_focus(test, props, data_fn):
         cells_are_same_width(t1, t2)
 
     assert test.get_log_errors() == []
-
-
-@pytest.mark.parametrize("props", basic_modes)
-@pytest.mark.parametrize(
-    "data_fn",
-    [generate_mock_data],
-)
-def test_szng004_on_focus(test, props, data_fn):
-    on_focus(test, props, data_fn)
-
-
-@pytest.mark.parametrize("props", basic_modes)
-@pytest.mark.parametrize(
-    "data_fn",
-    [generate_markdown_mock_data],
-)
-def test_szng005_on_focus(test, props, data_fn):
-    on_focus(test, props, data_fn)
-
-
-@pytest.mark.parametrize("props", basic_modes)
-@pytest.mark.parametrize(
-    "data_fn",
-    [generate_mixed_markdown_data],
-)
-def test_szng006_on_focus(test, props, data_fn):
-    on_focus(test, props, data_fn)
